Split ia2.h into its own shared library
ia2.h had code that didn't need to be duplicated in every compilation
unit, which was the case since functions were all declare static inline
in the header. This change starts to split that apart by moving
protect_pages into a shared library.
