+++ libia2/include/scrub_registers.h
